Add MyObject class with X and Y velocity
I replaced the BobObj class with a more generic MyObject class. This class
uses a "builder pattern" to construct the object, instead of multiple
constructors with different parameter sets.

In addition, I changed velocity to have both an X and a Y component. Now
the slow Bob moves diagonally instead of just straight across the screen.
Wrapping is still done at arbitrary locations, instead of doing collision
detection with the actual edges of the display.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>