Add "blue queen" bitmap
I found a neat picture of a playing card online; it is a
blue queen of clubs <1>. I created a playing card bitmap
from it using my non-existant GIMP skills and added it to the
project. It is now the third bitmap in AnimatedBitmapObjects.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>

<1> Here is the URL where I found the image; I do not intend to
use it for anything other than my own learning purposes so I
don't think I need to do anything other than provide attribution.

http://enjoyinspiration.com/graphic-inspiration-mm-brand-agency-playing-card-calendar/