Refactor MyObject to use explicit size instead of scale
Specifying the scaling factor seemed logical at first, but once
I started drawing bitmaps it became clear that the scaling is
done automatically by the draw methods, based on the destination
Rectangle provided. Defining the rectangle is easiest when the
object knows its size. The coder almost certainly knows the absolute
size of the object being created, and it has nothing to do with the
size of the skin, but is instead based on the contents of the display.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>