Add first touch interface to Play Deck 1
The class listens for touch events. When the draw pile is
touched, it is marked selected (using the new interface).
When the touch is removed, it is marked not selected.

This implementation is crude. In the future, the objects
will be tracked in a list and handled anonymously. After
that, dragging will be implemented.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>