Add ability to move the cards around
A MyObject can now be moved by a new updatePosition() method
that takes two parameters, the delta-X and delta-Y to apply.
The render thread records the location of each UP and MOVE
motion and uses it to provide the delta. This results in the
select card being moved around as the user directs.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>