Improve placement of cards in Swipe sample
I did the math to figure out how to place the cards without using
static sizes for the display area or the cards. Added some methods
to MyObject to allow the size to be modified, including with scaling
so that the dimensions stay the same. Having these new methods prompted
me to create a unit test suite for MyObject.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>