Add falling ball
A red ball is created at the top of the play area and it begins falling
immediately with a constant velocity. If the ball hits the paddle, the ball
will be placed back at the top of the play area and will begin to fall again.
If the ball is missed by the paddle, the ball will be placed at the top center
of the play area but it will no longer move.

This obviously isn't a game per-se, but it at least demonstrates how a ball
and paddle can interact.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>