Add simple gravity effect
The ground now applies an acceleration to the ball. This effect is applied
to the velocity of the ball before its new position is calculated. The code
is simple; in the future I will make it more generic, probably by having a list
of objects that apply acceleration and a list of objects that are affected by
acceleration.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>