Implement simple gravity method in MyObject
My thought is to have gravity be a part of MyObject. This will involve
having gravity as a data member (many objects will have this be 0) as well
as a method to apply gravity from a source object to the current object. This
code does this in a very simple way as a proof of concept. It needs to be
expanded to handle point sources and line sources.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>