Add paddle velocity and walls
If the paddle is moving when it hits the ball, the ball is given a bit
of velocity in the direction that the paddle was moving. The velocity
calculation does not seem quite right; it comes out 0 even when the paddle
is obviously moving. I think this has to do with how the run() thread is
scheduled in relation to the onTouchEvent() thread, but I was not able to
make a complete determination of this. On the emulator, there seemed to be
several touch events in between each run, while on my phone they alternated.

Walls have been added to the play area. The ball simply bounces off them
with no change to the magnitude of the velocity. The walls have been made
very wide so that collision detection will always work. In the future, the
collision detection algorithm should be improved so that this is not
necessary. One possibility is to check for intersection with the line
created by the previous position of the ball with its new position. If the
line crosses any edge of the wall then a collision has occurred.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>