Add plobjects library
This is my first attempt at using an AAR library module. It seems to have
worked as expected. The importation was not difficult, and I was able to
create an object (ball2) based on the PLObject class. It doesn't do anything,
but the point of this commit was to get the library imported.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>