Replace MyObject with PLObject
With this commit, MyObject is no longer used in Falling Down.
Instead, all objects are of type PLObject. Functionality is the
same as it was before the conversion to PLObject began: a single
ball bounces around the arena via a paddle controlled by the player.
The only difference is that there is no change to the velocity of the
ball based on the velocity of the paddle at the moment of contact.
In the future, the balls trajectory will be changed based on how far
from the center of the paddle it hit, which is apparently the common
way to do it.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>